Added TableGenerator, extension methods for using it, and tests
The TableGenerator class was created to make generating html tables
easier. To use it columns must be specified by adding them using
either model expressions or model transformation functions. The table
may either have no header,an empty header for a column, a textual
header, or another html tag as a header. Headers are specified as
the columns definitions of the table are added. Optionally column
tranformation functions can be provided that can be used to modify
each column as it is generated. Also, the header row, and/or each
body row may be modified as they are generated by providing
transformation functions. It is also possible to generate a table
without a header entirely.

Two extension methods were also provided for use in ASP pages that
can be used with any asp page that implements IFubuPage<TModel>.
Using these extension methods a table could be generated on an
ASP page like this:

<%= this.TableFor(m => m.Products, p => p.Code, p => p.Name, p => p.Parts) %>

or, if more control is needed (in this case a class added to the header row):

<%= this.TableGeneratorFor(m => m.Products, p => p.Code, p => p.Name)
        .AddHeaderRowTransformer(r => r.AddClass("header"))
        .GenerateTable(Model.Products) %>

It is also worth noting that, if a column is added using a model
expression, as the columns are generated the value evaluated from
using the expression (for each column in each row) will be run
through the ITableGenerator.DisplayFor() mechanism.